pyproject.toml
[tool.poetry]
name = "gcal_sync"
version = "0.1.0"
description = ""
authors = ["Yasunori Horikoshi <hotoku@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
google-api-python-client = "^2.68.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.7.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
